backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
playwright==1.44.0
openai==1.33.0
beautifulsoup4==4.12.3
pypdf==4.2.0
faiss-cpu==1.8.0
supabase==2.5.0
numpy==1.26.4
python-multipart==0.0.9
requests==2.32.3
